chore: generic 7.0.0 usage, regex engine moved to generic, @Discoverable usage for providing data to the about page
