+++ src/main/java/ch/sbb/polarion/extension/pdf/exporter/rest/controller/ConfigurationApiController.java
@@ -1,8 +1,8 @@
 package ch.sbb.polarion.extension.pdf.exporter.rest.controller;
 
+import ch.sbb.polarion.extension.generic.configuration.ConfigurationStatus;
 import ch.sbb.polarion.extension.generic.rest.filter.Secured;
 import ch.sbb.polarion.extension.generic.service.PolarionService;
-import ch.sbb.polarion.extension.pdf.exporter.rest.model.configuration.ConfigurationStatus;
 import org.jetbrains.annotations.NotNull;
 
 import javax.ws.rs.Path;
