+++ src/main/java/ch/sbb/polarion/extension/pdf/exporter/rest/controller/ConfigurationInternalController.java
@@ -1,7 +1,13 @@
 package ch.sbb.polarion.extension.pdf.exporter.rest.controller;
 
-import ch.sbb.polarion.extension.pdf.exporter.rest.model.configuration.ConfigurationStatus;
-import ch.sbb.polarion.extension.pdf.exporter.util.configuration.ConfigurationStatusUtils;
+import ch.sbb.polarion.extension.generic.configuration.ConfigurationStatus;
+import ch.sbb.polarion.extension.generic.configuration.ConfigurationStatusProvider;
+import ch.sbb.polarion.extension.pdf.exporter.util.configuration.CORSStatusProvider;
+import ch.sbb.polarion.extension.pdf.exporter.util.configuration.DefaultSettingsStatusProvider;
+import ch.sbb.polarion.extension.pdf.exporter.util.configuration.DleToolbarStatusProvider;
+import ch.sbb.polarion.extension.pdf.exporter.util.configuration.DocumentPropertiesPaneStatusProvider;
+import ch.sbb.polarion.extension.pdf.exporter.util.configuration.LiveReportMainHeadStatusProvider;
+import ch.sbb.polarion.extension.pdf.exporter.util.configuration.WeasyPrintStatusProvider;
 import io.swagger.v3.oas.annotations.Hidden;
 import io.swagger.v3.oas.annotations.Operation;
 import org.jetbrains.annotations.NotNull;
@@ -23,7 +29,7 @@ public class ConfigurationInternalController {
     @Path("/configuration/default-settings")
     @Operation(summary = "Checks default settings configuration")
     public @NotNull ConfigurationStatus checkDefaultSettings(@QueryParam("scope") @DefaultValue("") String scope) {
-        return ConfigurationStatusUtils.getSettingsStatus(scope);
+        return new DefaultSettingsStatusProvider().getStatus(ConfigurationStatusProvider.Context.builder().scope(scope).build());
     }
 
     @GET
@@ -31,7 +37,7 @@ public class ConfigurationInternalController {
     @Path("/configuration/document-properties-pane-config")
     @Operation(summary = "Checks document properties pane configuration")
     public @NotNull ConfigurationStatus checkDocumentPropertiesPaneConfig(@QueryParam("scope") @DefaultValue("") String scope) {
-        return ConfigurationStatusUtils.getDocumentPropertiesPaneStatus(scope);
+        return new DocumentPropertiesPaneStatusProvider().getStatus(ConfigurationStatusProvider.Context.builder().scope(scope).build());
     }
 
     @GET
@@ -39,7 +45,7 @@ public class ConfigurationInternalController {
     @Path("/configuration/dle-toolbar-config")
     @Operation(summary = "Checks DLE Toolbar configuration")
     public @NotNull ConfigurationStatus checkDleToolbarConfig() {
-        return ConfigurationStatusUtils.getDleToolbarStatus();
+        return new DleToolbarStatusProvider().getStatus(ConfigurationStatusProvider.Context.builder().build());
     }
 
     @GET
@@ -47,7 +53,7 @@ public class ConfigurationInternalController {
     @Path("/configuration/live-report-config")
     @Operation(summary = "Checks Live Report configuration")
     public @NotNull ConfigurationStatus checkLiveReportConfig() {
-        return ConfigurationStatusUtils.getLiveReportMainHeadStatus();
+        return new LiveReportMainHeadStatusProvider().getStatus(ConfigurationStatusProvider.Context.builder().build());
     }
 
     @GET
@@ -55,7 +61,7 @@ public class ConfigurationInternalController {
     @Path("/configuration/cors-config")
     @Operation(summary = "Checks CORS configuration")
     public @NotNull ConfigurationStatus checkCORSConfig() {
-        return ConfigurationStatusUtils.getCORSStatus();
+        return new CORSStatusProvider().getStatus(ConfigurationStatusProvider.Context.builder().build());
     }
 
     @GET
@@ -63,6 +69,6 @@ public class ConfigurationInternalController {
     @Path("/configuration/weasyprint")
     @Operation(summary = "Checks WeasyPrint configuration")
     public @NotNull List<ConfigurationStatus> checkWeasyPrint() {
-        return ConfigurationStatusUtils.getWeasyPrintStatus();
+        return new WeasyPrintStatusProvider().getStatuses(ConfigurationStatusProvider.Context.builder().build());
     }
 }
