+++ src/main/java/ch/sbb/polarion/extension/pdf/exporter/util/LiveDocCommentsProcessor.java
@@ -1,7 +1,7 @@
 package ch.sbb.polarion.extension.pdf.exporter.util;
 
 import ch.sbb.polarion.extension.pdf.exporter.model.LiveDocComment;
-import ch.sbb.polarion.extension.pdf.exporter.util.regex.RegexMatcher;
+import ch.sbb.polarion.extension.generic.regex.RegexMatcher;
 import com.polarion.alm.server.api.model.document.ProxyDocument;
 import com.polarion.alm.shared.api.model.comment.CommentBase;
 import com.polarion.alm.shared.api.model.comment.CommentBaseFields;
